Add DirLight

+++ OpenMC/Src/Helpers/UtilTool.h
@@ -4,7 +4,7 @@
 class UtilTool
 {
 public:
-  static float scaleValue(float value) {
+  static double scaleValue(float value) {
     return sin((value * 3.1415) - 1.57) * 0.5 + 0.5;
   }
